Create cibersortx_plots.R
This R script creates bar plots for every cell type, every sample, and all the samples combined in one plot from the CIBERSORTx results. The only thing that needs to be changed upon use is line 4 where the user must update the result file name. It uses the ggplot2 package to create the graphs and there is some minor manipulation of rows and columns necessary for the creation of the plots executed by this code.